refactor: LotteryMachine에 LotteryGenerator 기능을 합침
- LotteryMachine과 LotteryGenerator에 같은 이름의 메서드가 있고, LotteryMachine 클래스의 이름에 LotteryGenerator 클래스의 역할이 포함된다고 판단하여 합쳤음
