+++ docs/README.md
@@ -5,8 +5,9 @@
 ## 기능 목록 작성
 
 - LotteryMachine
-- [x] 당첨 번호와 로또들을 받아 결과를 알려준다.
-- [x] 구매 금액을 받아 금액에 맞는 개수의 로또를 발행한다.
+- [x] 수동 로또와 자동 로또 개수를 받아 로또들을 발행한다. 
+- [x] 발행 개수를 받아 랜덤으로 로또들을 발행한다.
+- [x] 랜덤으로 로또를 발행한다.
 
 - PurchaseAmount
 - [x] 구입금액을 가진다.
