--- src/main/kotlin/lotto/domain/LotteryGenerator.kt
@@ -1,20 +1,0 @@
-package lotto.domain
-
-import lotto.domain.LotteryNumber.Companion.UPPER_BOUNDARY
-
-class LotteryGenerator {
-
-    fun generateLotteries(count: Int): Lotteries {
-        val randomLotteries: MutableList<Lottery> = mutableListOf()
-        repeat(count) { randomLotteries.add(generateLottery()) }
-
-        return Lotteries(randomLotteries)
-    }
-
-    private fun generateLottery(): Lottery {
-        val randomLotteryNumberCandidates: List<LotteryNumber> =
-            List(UPPER_BOUNDARY) { i -> LotteryNumber(i + 1) }.shuffled()
-
-        return Lottery(randomLotteryNumberCandidates.subList(0, Lottery.NUMBER_SIZE))
-    }
-}
