+++ src/main/kotlin/lotto/domain/LotteryMachine.kt
@@ -1,13 +1,21 @@
 package lotto.domain
 
-class LotteryMachine(
-    private val generator: LotteryGenerator = LotteryGenerator()
-) {
+class LotteryMachine() {
 
     fun createLotteries(lotteries: Lotteries, autoNumber: Int): Lotteries =
-        lotteries.plus(generateLotteries(autoNumber))
+        lotteries.plus(generateRandomLotteries(autoNumber))
 
-    fun generateLotteries(count: Int): Lotteries {
-        return generator.generateLotteries(count)
+    fun generateRandomLotteries(count: Int): Lotteries {
+        val randomLotteries: MutableList<Lottery> = mutableListOf()
+        repeat(count) { randomLotteries.add(generateRandomLottery()) }
+
+        return Lotteries(randomLotteries)
+    }
+
+    private fun generateRandomLottery(): Lottery {
+        val randomLotteryNumberCandidates: List<LotteryNumber> =
+            List(LotteryNumber.UPPER_BOUNDARY) { i -> LotteryNumber(i + 1) }.shuffled()
+
+        return Lottery(randomLotteryNumberCandidates.subList(0, Lottery.NUMBER_SIZE))
     }
 }
